char/sdma: Support multiple sdma devices
There are multiple sdma devices on board that contains multiple Parts.
We supply an additional virtual device in /dev/sdma and the real sdma
devices are named /dev/sdmaX where X is the index. When the user operate
on /dev/sdma we choose a proper device for user.

Currently the index of the sdma device is defined according to the numa
node it belongs to, so we select the device depending on the numa node
of the cpus the user runs on. The device was decided during the open
operation and would not change.

Signed-off-by: Wang Wensheng <wangwensheng4@huawei.com>
Signed-off-by: Zhang Zekun <zhangzekun11@huawei.com>
Reviewed-by: Weilong Chen <chenweilong@huawei.com>