driver/svm: support SVSP feature, interface definition
ascend inclusion
category: bugfix
bugzilla: NA
DTS: NA
CVE: NA

-----------------------------------------------------------------

In the SVA scenario, the master device of the SMMU can access the
address space of the process by the operation of binding proces.
The page table of the process is shared by the MMU and SMMU, In
this case, if some addrss range cannot be accessed through the MMU,
access to this address range needs to be isolated from the process
page table.

If the default domain of the SMMU is used, the isolation can be
achieved. However, the address space created by default domain do
not like the address space of the process which can be shared by
multiple SMMUs.

Therefore, we add a shadow virtual space of process. This address
space is used only by the SMMU.

Signed-off-by: Zhou Guanghui <zhouguanghui1@huawei.com>
Signed-off-by: Zhang Zekun <zhangzekun11@huawei.com>
Reviewed-by: Weilong Chen <chenweilong@huawei.com>