driver/iommu: Introduce IOMMU_SVA_FEAT_SVSP feature
ascend inclusion
category: feature
bugzilla: NA
CVE: NA

-----------------------------------------------------------------

The svsp requires the pasid of svsp_mm has the value of pasid of
the normal mm with the higest bit set.
In this commit we introduce IOMMU_SVA_FEAT_SVSP feature to allow
user pass the pasid of normal mm and set pasid of svsp_mm by the
given rules.

Signed-off-by: Yuan Can <yuancan@huawei.com>
Signed-off-by: Zhang Zekun <zhangzekun11@huawei.com>
Reviewed-by: Weilong Chen <chenweilong@huawei.com>